+++ geopandas/array.py
@@ -552,6 +552,13 @@ class GeometryArray(ExtensionArray):
         return shapely.has_z(self._data)
 
     @property
+    def has_m(self):
+        if not SHAPELY_GE_21:
+            raise ImportError("'has_m' requires shapely>=2.1.")
+
+        return shapely.has_m(self._data)
+
+    @property
     def geom_type(self):
         res = shapely.get_type_id(self._data)
         return geometry_type_values[np.searchsorted(geometry_type_ids, res)]
@@ -1236,6 +1243,25 @@ class GeometryArray(ExtensionArray):
                 return shapely.get_z(self._data)
         else:
             message = "z attribute access only provided for Point geometries"
+            raise ValueError(message)
+
+    @property
+    def m(self):
+        """Return the m coordinate of point geometries in a GeoSeries"""
+        if not SHAPELY_GE_21:
+            raise ImportError("'m' requires shapely>=2.1.")
+
+        if (self.geom_type[~self.isna()] == "Point").all():
+            empty = self.is_empty
+            if empty.any():
+                nonempty = ~empty
+                coords = np.full_like(nonempty, dtype=float, fill_value=np.nan)
+                coords[nonempty] = shapely.get_m(self._data[nonempty])
+                return coords
+            else:
+                return shapely.get_m(self._data)
+        else:
+            message = "m attribute access only provided for Point geometries"
             raise ValueError(message)
 
     @property